move the global setup file to the root level
