+++ .gitignore
@@ -38,3 +38,4 @@ coverage
 /playwright/.cache/
 /playwright/.auth/
 /.auth/*
+storageState.json
